avalanche integration refactoring
