+++ src/renate/updaters/model_updater.py
@@ -199,6 +199,7 @@ class ModelUpdater(abc.ABC):
             )
             early_stopping_enabled = False
 
+        self._current_state_folder = current_state_folder
         self._next_state_folder = next_state_folder
         self._metric = metric
         self._mode = mode
@@ -220,10 +221,6 @@ class ModelUpdater(abc.ABC):
             self._transforms_kwargs["buffer_transform"] = self._buffer_transform
             self._transforms_kwargs["buffer_target_transform"] = self._buffer_target_transform
         self._max_epochs = max_epochs
-        self._learner = self._load_learner(learner_class, self._learner_kwargs)
-        assert self._learner.is_logged_metric(metric), f"Target metric `{metric}` is not logged."
-        self._logger = logger
-        self._num_epochs_trained = 0
         if accelerator not in defaults.SUPPORTED_ACCELERATORS:
             raise ValueError(
                 f"Accelerator {accelerator} not supported. "
@@ -231,6 +228,10 @@ class ModelUpdater(abc.ABC):
             )
         self._accelerator = accelerator
         self._devices = devices
+        self._learner = self._load_learner(learner_class, self._learner_kwargs)
+        assert self._learner.is_logged_metric(metric), f"Target metric `{metric}` is not logged."
+        self._logger = logger
+        self._num_epochs_trained = 0
 
     @abc.abstractmethod
     def update(
